requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.3.1
django-anymail==13.0
django-cors-headers==4.6.0
django-location-field==2.7.3
django-oauth-toolkit==3.0.1
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-social-oauth2==3.1.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
paystackapi==2.1.3
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.11.0
sentry-sdk==2.19.2
setuptools==75.8.2
sib-api-v3-sdk==7.6.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
twilio==9.4.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.18.3
